+++ app/Http/Resources/TeamMemberResource.php
@@ -2,26 +2,27 @@
 
 namespace App\Http\Resources;
 
+use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
+use JsonSerializable;
 
 class TeamMemberResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
+     * @param Request $request
+     * @return array|Arrayable|JsonSerializable
      */
-  public function toArray($request)
-  {
-    $user = $this->whenLoaded('user');
+  public function toArray(Request $request): array|JsonSerializable|Arrayable {
     return [
         'id' => $this->id,
-        'name' => $user->name ?? '',
-        'email' => $user->email ?? '',
-        'phone' => $user->phone ?? '',
-        'profile_photo_path' => $user->profile_photo_path ?? '',
-        'profile_photo_url' => $user->profile_photo_url ?? '',
+        'name' => $this->name ?? '',
+        'email' => $this->email ?? '',
+        'phone' => $this->phone ?? '',
+        'profile_photo_path' => $this->profile_photo_path ?? '',
+        'profile_photo_url' => $this->profile_photo_url ?? '',
         'active' => $this->pivot->active ?? null,
         'team_profile_is_public' => $this->pivot->team_profile_is_public ?? null,
         'created_at' => $this->pivot->created_at ?? null,